Update Couchbase versions
Remove Couchbase version 6.6, which was EOL in Oct 2023.

Add version 7.2 and 7.6, EOL in Jul 2026 and Mar 2027, respectively.

Keep version 7.0 and 7.1, even though they were EOL in July 2025. This
keeps some backwards compatibility, and also reflects Couchbase's own
documentation pages, where 7.0 and 7.1 are still available to browse.

Discontinue the cbdocloader, which is not available in version 7.6.
Instead, use the REST endpoint /sampleBuckets/install.

Remove the expected index stat tag "bucket:cb_bucket", as it is not
present in the index stats for either version. It's not clear how that
ever worked. Instead, rely on the "bucket:gamesim-sample" tag.